Allow hooks to declare themselves "exclusive".
If a hook is exclusive, then it is only run when the pagetype of the
file is the same as the id of the hook.  This should especially speed
up the execution of ikiwiki when using external plugins, since much of
the time spent in external plugins is moving content into the plugin
before it checks to see if should actually operate on the content.

It is not clear to me whether the run_hooks function has enough
information to determine the page type in every situation, so I added
that as an optional argument, and filled in the argument in situations
where my plugin would see a speedup.  If run_hooks can determine the
page type in these situations, then it probably should, and the extra
argument should be gotten rid of.